mcp-server/pyproject.toml
[project]
name = "finance-mcp-server"
version = "0.1.0"
description = "Finance MCP Server for Portfolio Analysis"
authors = [
    {name = "Finance App", email = "finance@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "beautifulsoup4>=4.13.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false  # This is an application, not a library

[project.scripts]
finance-mcp = "finance_mcp_server:main"
generic-mcp = "generic_mcp_server:main"
